requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
numpy==1.24.3
python-binance==1.0.19
schedule==1.2.0 